+++ pygmo/s_policy.cpp
@@ -7,25 +7,20 @@
 // with this file, You can obtain one at http://mozilla.org/MPL/2.0/.
 
 #include <memory>
-#include <sstream>
 #include <string>
 #include <typeindex>
 #include <typeinfo>
-#include <vector>
-
-#include <boost/numeric/conversion/cast.hpp>
 
 #include <pybind11/numpy.h>
 #include <pybind11/pybind11.h>
 
 #include <pagmo/config.hpp>
-#include <pagmo/s11n.hpp>
 #include <pagmo/s_policy.hpp>
 #include <pagmo/types.hpp>
 
 #include "common_utils.hpp"
 #include "handle_thread_py_exception.hpp"
-#include "object_serialization.hpp"
+#include "s11n_wrappers.hpp"
 #include "s_policy.hpp"
 
 namespace pagmo
@@ -129,17 +124,13 @@ void *s_pol_inner<py::object>::get_ptr()
 template <typename Archive>
 void s_pol_inner<py::object>::save(Archive &ar, unsigned) const
 {
-    ar << boost::serialization::base_object<s_pol_inner_base>(*this);
-    ar << pygmo::object_to_vchar(m_value);
+    pygmo::inner_class_save<s_pol_inner_base>(ar, *this);
 }
 
 template <typename Archive>
 void s_pol_inner<py::object>::load(Archive &ar, unsigned)
 {
-    ar >> boost::serialization::base_object<s_pol_inner_base>(*this);
-    std::vector<char> v;
-    ar >> v;
-    m_value = pygmo::vchar_to_object(v);
+    pygmo::inner_class_load<s_pol_inner_base>(ar, *this);
 }
 
 } // namespace detail
@@ -147,53 +138,3 @@ void s_pol_inner<py::object>::load(Archive &ar, unsigned)
 } // namespace pagmo
 
 PAGMO_S11N_S_POLICY_IMPLEMENT(pybind11::object)
-
-namespace pygmo
-{
-
-namespace py = pybind11;
-
-// Serialization support for the s_policy class.
-py::tuple s_policy_pickle_getstate(const pagmo::s_policy &s)
-{
-    // The idea here is that first we extract a char array
-    // into which s has been serialized, then we turn
-    // this object into a Python bytes object and return that.
-    std::ostringstream oss;
-    {
-        boost::archive::binary_oarchive oarchive(oss);
-        oarchive << s;
-    }
-    auto str = oss.str();
-    return py::make_tuple(py::bytes(str.data(), boost::numeric_cast<py::size_t>(str.size())));
-}
-
-pagmo::s_policy s_policy_pickle_setstate(py::tuple state)
-{
-    // Similarly, first we extract a bytes object from the Python state,
-    // and then we build a C++ string from it. The string is then used
-    // to deserialized the object.
-    if (py::len(state) != 1) {
-        pygmo::py_throw(PyExc_ValueError, ("the state tuple passed for selection policy deserialization "
-                                           "must have 1 element, but instead it has "
-                                           + std::to_string(py::len(state)) + " element(s)")
-                                              .c_str());
-    }
-
-    auto ptr = PyBytes_AsString(state[0].ptr());
-    if (!ptr) {
-        pygmo::py_throw(PyExc_TypeError, "a bytes object is needed to deserialize a selection policy");
-    }
-
-    std::istringstream iss;
-    iss.str(std::string(ptr, ptr + py::len(state[0])));
-    pagmo::s_policy s;
-    {
-        boost::archive::binary_iarchive iarchive(iss);
-        iarchive >> s;
-    }
-
-    return s;
-}
-
-} // namespace pygmo