+++ pygmo/topology.hpp
@@ -79,14 +79,4 @@ struct topo_inner<py::object> final : topo_inner_base, pygmo::common_base {
 // Register the topo_inner specialisation for py::object.
 PAGMO_S11N_TOPOLOGY_EXPORT_KEY(pybind11::object)
 
-namespace pygmo
-{
-
-namespace py = pybind11;
-
-py::tuple topology_pickle_getstate(const pagmo::topology &);
-pagmo::topology topology_pickle_setstate(py::tuple);
-
-} // namespace pygmo
-
 #endif